requirements.txt
# Core trading dependencies
alpaca-py==0.20.0
pandas==2.1.4
numpy==1.24.4
pydantic==2.5.2
PyYAML==6.0.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Technical analysis
talib==0.4.28
vectorbt==0.26.0
yfinance==0.2.18

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
pytz==2023.3
schedule==1.2.0

# Monitoring
prometheus-client==0.19.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Optional notification
slack-sdk==3.26.1